### 2021-10-04 #3789 @danielskatz
@chartgerink - ok, thanks - @kyleniemeyer will look for another editor. Also, you might want to change your max papers on the JOSS dashboard from 4 to 3 to help AEiCs.
