### 2021-11-09 #3901 @whedon
[REVIEW]: bistablehistory: an R package for history-dependent analysis of perceptual time series
**Submitting author:** @alexander-pastukhov (<a href="http://orcid.org/0000-0002-8738-8591">Alexander Pastukhov</a>)
**Repository:** <a href="https://github.com/alexander-pastukhov/bistablehistory" target ="_blank">https://github.com/alexander-pastukhov/bistablehistory</a>
**Version:** v1.1.0
