### 2021-11-10 alexander-pastukhov/bistablehistory#1 @alexander-pastukhov
Dear @fernandomayer,

thank you and terribly sorry, I was so happy that the CRAN version works that I did not double check the developmental version. I've added information on installation from CRAN to README (`install.packages("bistablehistory")` that one definitely works and the installation is much quicker with precompiled binaries). For the developmental version the exports were indeed MIA somehow, not sure how that could have happened given that they are generated by roxygen2. I've re-roxygenized it, so the exports are there and package installs correctly and works.

Line #9   `Warning in system2(command, args, stdout = NULL, stderr = NULL, ...) :     error in running command`
Cannot replicate it on Windows but it is looks like a part of the general package infrastructure.
