### 2021-11-10 alexander-pastukhov/bistablehistory#1 @fernandomayer
Thanks.

I tested both the CRAN and the Github versions and now both are working.

Just a few additional comments:

- The line #9 warning is still there when installing from Github, but that didn't affect the installation. I suppose this is a minor issue, but maybe it could be addressed in the future. If it passed the CRAN checks, it possibly is an issue specific to my system.
- Precompiled binaries are available only for Windows. Any package installed from CRAN on Linux will be compiled from source anyway (unless you explicitly provide precompiled binaries for Linux, which is very uncommon - and I think it is even not recommended). This is not a problem *per se*, just to let you know.
- Maybe it would be useful to inform in `README` that the Github version in the `master` branch is the development version (it may not be obvious to everyone)
