### 2021-11-11 alexander-pastukhov/bistablehistory#1 @alexander-pastukhov
* line #9, no warnings or even notes on that, even during the official CRAN check (that tends to catch more things than one in devtools).
* Not sure whether I can do anything about this as this is on CRAN's side, I only submit source package (as recommended).
* Thank you, good point! Added this info to README.
