### 2021-12-20 #3901 @fartist
I plan to finish my review in the coming week, thx! Sent from Mail for Windows From: Mikkel Meyer AndersenSent: Friday, December 17, 2021 1:29 AMTo: openjournals/joss-reviewsCc: Feeeeeeeeeeng; MentionSubject: Re: [openjournals/joss-reviews] [REVIEW]: bistablehistory: an R package for history-dependent analysis of perceptual time series (Issue ***@***.***, @fartist, can you please give a brief status of your review? This is not to rush you, merely to give me an impression of the progress and time-frame.—Reply to this email directly, view it on GitHub, or unsubscribe.Triage notifications on the go with GitHub Mobile for iOS or Android. You are receiving this because you were mentioned.Message ID: ***@***.***>
