### 2021-12-27 alexander-pastukhov/bistablehistory#3 @fartist
[REVIEW]joss
1. line 16/17 how the approach violates the iid assumption is not clear to me, I would appreciate more explanation on this part.
2. line 18/19: although it cites the prior work on the homogenous fist order process, I would believe that including an explicit model formulation should be easier for readers to see roughly how the underlying model works. Especially in your case, you used Stan, a full Bayesian procedure, so including a simple description for your likelihood function and prior specification would be helpful for users. Or maybe I could suggest linking your vignette page (e.g., https://cran.rstudio.com/web/packages/bistablehistory/vignettes/cumulative-history.html). btw, although it is not needed for the sake of publishing this paper, a better documentation for the model formulation would be more useful (e.g.,
![image](https://user-images.githubusercontent.com/14242695/147424267-c7c28922-03f2-46ed-9056-228dc9010763.png)
in your vignette, $A$ and $t^prime$, for example, are not defined.
3. line 26/27: "thus it provides .... ability to compare models via information criteria, etc", this sounds vague to me as frequentist likelihood inference can also do model comparisons using information criteria (e.g., AIC) so maybe better to say explicitly you are referring here WAIC and/or DIC.
4. Is there any current R packages or any other software packages for history-dependent analysis? if so, it would be good to cite.
