### 2021-12-27 #3901 @fartist
Please find my review at https://github.com/alexander-pastukhov/bistablehistory/issues/3 - solid work!
