### 2022-01-16 #3901 @alexander-pastukhov
@fartist  I have posted detailed response at [alexander-pastukhov/bistablehistory#3](https://github.com/alexander-pastukhov/bistablehistory/issues/3) , should I duplicate it here as well?
