### 2022-01-16 #3901 @fartist
No, thanks - I can see that on the repo - thanks for taking your time.

On Sun, Jan 16, 2022 at 3:46 AM Alexander (Sasha) Pastukhov <
***@***.***> wrote:

> @fartist <https://github.com/fartist> I have posted detailed response at
> alexander-pastukhov/bistablehistory#3
> <https://github.com/alexander-pastukhov/bistablehistory/issues/3> ,
> should I duplicate it here as well?
>
