### 2022-02-04 #3901 @mikldk
@alexander-pastukhov Is it correct that you have addressed all points in alexander-pastukhov/bistablehistory#3? In that case I guess we just need @fartist to update the review.
