### 2022-02-16 #3901 @alexander-pastukhov
@mikldk

Version 1.1.0
Tag v1.1.0
https://github.com/alexander-pastukhov/bistablehistory/releases/tag/v1.1.0

Zenodo DOI for 1.1.0: https://zenodo.org/record/5851674
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5851674.svg)](https://doi.org/10.5281/zenodo.5851674)
