### 2022-02-16 #3901 @mikldk
> @mikldk
>
> Version 1.1.0 Tag v1.1.0 https://github.com/alexander-pastukhov/bistablehistory/releases/tag/v1.1.0
>
> Zenodo DOI for 1.1.0: https://zenodo.org/record/5851674 [![DOI](https://camo.githubusercontent.com/8d7b2993811bdf13bafbbb2f51561dcc4e88c3187dab23f6755604c6d477194d/68747470733a2f2f7a656e6f646f2e6f72672f62616467652f444f492f31302e353238312f7a656e6f646f2e353835313637342e737667)](https://doi.org/10.5281/zenodo.5851674)

@alexander-pastukhov , Please change the metadata at Zenodo (title, authors) to match the paper's.
