### 2022-02-17 #3901 @alexander-pastukhov
My bad, somehow two papers were included twice in the bib file. I guess that was the problem,
