### 2022-02-17 #3901 @danielskatz
@alexander-pastukhov - I've suggested some edits in https://github.com/alexander-pastukhov/bistablehistory/pull/4. Please check this and merge it if they seem correct to you.
